requirements.txt
streamlit==1.15.1
tensorflow==2.11.0
Pillow==9.3.0
opencv-python==4.6.0.66